+++ back/controllers/AdminuserController.js
@@ -33,18 +33,23 @@ exports.adminShow = async (req, res) => {
 };
 
 exports.adminEditUser = async (req, res) => {
-    console.log("Je suis le controller Edit dans user", req.body);
-
+    console.log("Je suis le controller Edit dans user", req.body)
+    let firstname = req.body.firstname
+    let name = req.body.name
+    let image = req.file
     const user = await db.query(`SELECT * FROM user WHERE id = "${req.params.id}"`)
 
-    await db.query(`UPDATE user SET imguser = "${req.file.filename}",
-      firstname="${req.body.firstname}", name="${req.body.name}" WHERE id = "${req.params.id}"`);
-
-    const dir = path.join('./public/data/users');
-    deleteFile(dir, user[0].imguser)
-
-    await db.query(`UPDATE user SET imguser = '${req.file.filename}' WHERE id = "${req.params.id}"`)
-
+    if (firstname) {
+        await db.query(`UPDATE user SET firstname="${req.body.firstname}" WHERE id = "${req.params.id}"`)
+    }
+    if (name) {
+        await db.query(`UPDATE user SET name="${req.body.name}" WHERE id = "${req.params.id}"`)
+    }
+    if (image) {
+        const dir = path.join('./public/data/users')
+        deleteFile(dir, user[0].imguser)
+        await db.query(`UPDATE user SET imguser = '${req.file.filename}' WHERE id = "${req.params.id}"`)
+    }
     res.redirect("/admin")
 };
